pyproject.toml
[tool.poetry]
name = "blankon-test"
version = "0.1.0"
description = ""
authors = ["Altair Rostam"]
readme = "README.md"
package-mode = false

[tool.ruff]
exclude = [".git", ".ruff_cache", ".tox", ".venv", "__pypackages__", "env", ".env"]
line-length = 100
indent-width = 4
target-version = "py311"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "UP"]
ignore = []
extend-select = ["F", "I", "UP"]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.isort]
case-sensitive = true

# [tool.pytest.ini_options]
# DJANGO_SETTINGS_MODULE = "core.settings"
# django_find_project = false
# python_files = "*_test.py"
# pythonpath = ["."]
# testpaths = ["todo"]
# filterwarnings = "ignore::DeprecationWarning:^(?!node\\.).*:"

[tool.poetry.dependencies]
python = "^3.12"
ruff = "^0.4.4"
testcontainers = "^4.4.0"
django = "^5.0.6"
djangorestframework = "^3.15.1"
psycopg = {extras = ["binary"], version = "^3.1.19"}
python-dotenv = "^1.0.1"
django-filter = "^24.2"
pytest = "^8.2.0"
pytest-django = "^4.8.0"
model-bakery = "^1.18.0"
pytest-xdist = "^3.6.1"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
